fix: emit string enum values for WorkSessionStatus and SettlementOutcome/Status
Admin endpoint was casting WorkSessionStatus to int before mapping, so
session.status arrived at the frontend as 0/1/2 while every other endpoint
emitted the JSON string names. Removing the (int) casts and changing
AdminSessionDto.Status from int to WorkSessionStatus aligns the admin
endpoint with global JsonStringEnumConverter behaviour.

Frontend types changed from numeric unions to string unions throughout:
WorkSessionStatus, SettlementOutcome, SettlementStatus, AdminSession.status.
OUTCOME_LABELS/STATUS_LABELS re-keyed to string keys. All template
comparisons (=== 1 / === 2) updated to named string values.

+++ backend/Services/AdminService.cs
@@ -84,7 +84,7 @@ public class AdminService(AppDbContext context, UserManager<User> userManager) :
                         {
                             ClockIn = s.ClockIn,
                             ClockOut = s.ClockOut,
-                            Status = (int)s.Status,
+                            Status = s.Status,
                             Hours = CalcSessionHours(s),
                             Breaks = s.Breaks
                                 .OrderBy(b => b.BreakStart)
@@ -460,7 +460,7 @@ public class AdminService(AppDbContext context, UserManager<User> userManager) :
                         {
                             ClockIn = s.ClockIn,
                             ClockOut = s.ClockOut,
-                            Status = (int)s.Status,
+                            Status = s.Status,
                             Hours = CalcSessionHours(s),
                             Breaks = s.Breaks
                                 .OrderBy(b => b.BreakStart)
